Movie Create - Edit

+++ PacificStore/Models/Genre.cs
@@ -8,7 +8,8 @@ namespace PacificStore.Models
 {
     public class Genre
     {
-        public int Id { get; set; }
+        public byte Id { get; set; }
+
         [Required]
         public string Name { get; set; }
     }
